pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "whereami"
version = "0.1.0"
description = "A lightweight GeoGuessr-style prototype backed by FastAPI."
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.115,<1",
  "python-dotenv>=1.0,<2",
  "uvicorn[standard]>=0.30,<1",
]

[tool.setuptools]
packages = ["backend"]
